requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
SQLAlchemy>=2.0.0
asyncpg>=0.27.0
alembic>=1.11
python-jose>=3.3.0
argon2-cffi>=21.3.0

pydantic>=1.10,<2
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pytest-cov>=4.0.0
psycopg2-binary>=2.9.0
email-validator>=2.0.0
python-multipart>=0.0.6
